Extract OrientationEvent listener to another class Remove parentLayout from the constructor of FullscreenVideoView
