+++ app/src/main/java/bg/devlabs/fullscreenvideoviewexample/ActionBarActivity.kt
@@ -27,7 +27,7 @@ class ActionBarActivity : AppCompatActivity() {
         // Change the ActionBar title
         supportActionBar?.title = getString(R.string.action_bar_activity)
         val videoPath = "http://clips.vorwaerts-gmbh.de/VfE_html5.mp4"
-        fullscreenVideoView.init(videoPath, parentLayout)
+        fullscreenVideoView.init(videoPath)
                 .exitFullscreenDrawable(R.drawable.ic_launcher_background)
                 .enterFullscreenDrawable(R.drawable.ic_media_pause)
                 .progressBarColor(R.color.colorAccent)
