+++ app/src/main/java/bg/devlabs/fullscreenvideoviewexample/NoActionBarActivity.kt
@@ -18,6 +18,6 @@ class NoActionBarActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_no_action_bar)
         val videoPath = "http://clips.vorwaerts-gmbh.de/VfE_html5.mp4"
-        fullscreenVideoView.init(videoPath, parentLayout)
+        fullscreenVideoView.init(videoPath)
     }
 }
