+++ app/src/main/java/bg/devlabs/fullscreenvideoviewexample/RegularActivity.kt
@@ -18,6 +18,6 @@ class RegularActivity : Activity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_regular)
         val videoPath = "http://clips.vorwaerts-gmbh.de/VfE_html5.mp4"
-        fullscreenVideoView.init(videoPath, parentLayout)
+        fullscreenVideoView.init(videoPath)
     }
 }
